# common/src/api/generated/ping/client.ts
/* Generated by npm run generate:api-contracts. Do not edit directly.
 * Source: api_spec/openapi_v2.yaml
 */

import { makeApi, Zodios, type ZodiosEndpointDefinitions, type ZodiosInstance, type ZodiosOptions } from "@zodios/core";
import { z } from "zod";

const ApiResponse = z
  .object({
    statusCode: z.number().int(),
    timestamp: z.string(),
    requestId: z.string(),
  })
  .passthrough();
const SuccessResponse = z
  .object({ statusCode: z.number().int(), content: z.object({}).passthrough() })
  .partial()
  .passthrough()
  .and(ApiResponse);

export const schemas: Record<string, z.ZodTypeAny> = {
  ApiResponse,
  SuccessResponse,
};

const endpoints: ZodiosEndpointDefinitions = makeApi([
  {
    method: "get",
    path: "/api/v2/ping",
    alias: "ping",
    description: `Returns a simple &#x27;pong&#x27; response. Useful for checking for connectivity to the server.
`,
    requestFormat: "json",
    response: SuccessResponse.and(
      z
        .object({
          statusCode: z.unknown(),
          content: z.object({ data: z.unknown() }).partial().passthrough(),
        })
        .partial()
        .passthrough()
    ),
  },
]);

export const pingApi: ZodiosInstance<typeof endpoints> = new Zodios(endpoints);

export function createApiClient(baseUrl: string, options?: ZodiosOptions): ZodiosInstance<typeof endpoints> {
  return new Zodios(baseUrl, endpoints, options);
}