# common/src/api/generated/subcontrollers/openapi.yaml
...
openapi: 3.0.1
info:
  title: sproot
  version: 0.3.0
  description: |
    Sproot is a general purpose environmental management system with a focus on greenhouse and garden. The system is intended to be used to monitor and control various environmental parameters, to  continuously update and maintain "homeostasis". This API is a machine-friendly interface for  interacting with the system, and is designed to be used by a web-based front-end. This specification can be used to generate client libraries, documentation, and other tools.
paths:
  /api/v2/subcontrollers:
    get:
      operationId: listSubcontrollers
      tags:
        - Subcontrollers
      summary: Get subcontrollers discovered via mDNS.
      description: |
        Returns a list of subcontrollers discovered on the local network via mDNS/Bonjour. This includes basic information about each device, such as name, host, and address.
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - properties:
                      statusCode:
                        example: 200
                      content:
                        properties:
                          data:
                            allOf:
                              - type: object
                                properties:
                                  recognizedDevices:
                                    type: array
                                    items:
                                      $ref: "#/components/schemas/SDBSubcontroller"
                                  unrecognizedDevices:
                                    type: array
                                    items:
                                      $ref: "#/components/schemas/UnrecognizedSubcontrollers"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - properties:
                      statusCode:
                        example: 500
                      error:
                        properties:
                          name:
                            example: Internal Server Error
                          url:
                            example: /api/v2/subcontrollers
                          details:
                            items:
                              example: Failed to retrieve subcontrollers.
    post:
      operationId: createSubcontroller
      tags:
        - Subcontrollers
      summary: Add a new subcontroller.
      description: |
        Adds a new subcontroller to the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubcontrollerCreateRequest"
      responses:
        "201":
          description: subcontroller created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - properties:
                      statusCode:
                        example: 201
                      content:
                        properties:
                          data:
                            $ref: "#/components/schemas/SDBSubcontroller"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - properties:
                      statusCode:
                        example: 400
                      error:
                        properties:
                          name:
                            example: Bad Request
                          url:
                            example: /api/v2/subcontrollers
                          details:
                            example:
                              - "Missing required field: name"
                              - "Missing required field: address"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - properties:
                      statusCode:
                        example: 500
                      error:
                        properties:
                          name:
                            example: Internal Server Error
                          url:
                            example: /api/v2/subcontrollers
                          details:
                            items:
                              example: Failed to add subcontroller.
...
components:
  schemas:
    SuccessResponse:
      description: |
        A success response object that is returned by all successful API requests.  Contains a status code, timestamp, a unique request Id, and, optionally, some content.
      allOf:
        - type: object
          properties:
            statusCode:
              type: integer
            content:
              type: object
              required:
                - data
        - $ref: "#/components/schemas/ApiResponse"
    SDBSubcontroller:
      description: |
        A subcontroller object that is stored in the database. Likely means an ESP32. Contains all the information needed to  reference a subcontroller, including name, ID, and host name.
      type: object
      required:
        - name
        - hostName
      properties:
        id:
          type: number
          readOnly: true
          example: 1
        name:
          type: string
          example: sproot-esp32-1C78
        hostName:
          type: string
          example: sproot-esp32-1C78.local
    UnrecognizedSubcontrollers:
      description: |
        An object representing a subcontroller discovered on the network.
      type: object
      required:
        - name
        - host
        - address
      properties:
        name:
          type: string
          example: sproot-esp32-1C78
        host:
          type: string
          example: sproot-esp32-1C78.local
        address:
          oneOf:
            - type: string
              example: 192.168.1.1
            - type: array
              items:
                type: string
    ErrorResponse:
      description: |
        An error response object that is returned by all unsuccessful API requests.  Contains a status code, timestamp, a unique request Id, and an error object. This error object may contain the name of the error, and an array of details (if relevant).
      allOf:
        - type: object
          required:
            - error
          properties:
            statusCode:
              type: integer
            error:
              type: object
              properties:
                request:
                  type: object
                  properties:
                    method:
                      type: string
                    url:
                      type: string
                    body:
                      type: object
                    query:
                      type: object
                name:
                  type: string
                url:
                  type: string
                details:
                  type: array
                  items:
                    type: string
        - $ref: "#/components/schemas/ApiResponse"
    SubcontrollerCreateRequest:
      description: |
        Request body for creating a subcontroller.
      type: object
      required:
        - name
        - hostName
      properties:
        name:
          type: string
          example: sproot-esp32-1C78
        hostName:
          type: string
          example: sproot-esp32-1C78.local
...
    ApiResponse:
      description: |
        A generic response object that is returned by all API endpoints. Contains a status code,  timestamp, and a unique request Id.
      type: object
      required:
        - statusCode
        - timestamp
        - requestId
      properties:
        statusCode:
          type: integer
        timestamp:
          type: string
          example: 2021-01-01T00:00:00.000Z
        requestId:
          type: string
          example: 9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d