# server/src/api/v2/ApiRootV2.ts
import { Express, Request, Response, NextFunction } from "express";
import { DI_KEYS } from "../../utils/DependencyInjectionConstants";
import * as OpenApiValidator from "express-openapi-validator";
import swaggerUi from "swagger-ui-express";
import YAML from "yamljs";

import addDefaultProperties, {
  createDefaultProperties,
} from "./middleware/DefaultResponseProperties";
import { authorize } from "./middleware/Authorize";
import authenticationRouter from "./authentication/AuthenticationRouter";
import pingRouter from "./ping/PingRouter";
import systemRouter from "./system/SystemRouter";
import sensorsRouter from "./sensors/SensorsRouter";
import outputsRouter from "./outputs/OutputsRouter";
import automationsRouter from "./automations/AutomationsRouter";
import outputActionsRouter from "./automations/OutputActionRouter";
import notificationsRouter from "./automations/NotificationActionRouter";
import cameraRouter from "./camera/CameraRouter";
import subcontrollersRouter from "./subcontrollers/SubcontrollersRouter";
import deviceZoneRouter from "./devicezones/DeviceZoneRouter";
import journalRouter from "./journals/JournalRouter";
import entryRouter from "./journals/EntriesRouter";
import tagRouter from "./tags/TagRouter";
import ContractValidationError from "../validation/ContractValidationError";

const spec_path = "../api_spec/openapi_v2.yaml";

const openapi_v2_doc = YAML.load(spec_path);
const swaggerUiOptions = {
  swaggerOptions: { defaultModelsExpandDepth: -1 },
};
const authenticateMiddleware = authorize(
  process.env["AUTHENTICATION_ENABLED"]!,
  process.env["JWT_SECRET"]!,
);

function ApiRootV2(app: Express) {
  const logger = app.get(DI_KEYS.Logger);

  app.use(
    "/api/v2/docs",
    swaggerUi.serveFiles(openapi_v2_doc, swaggerUiOptions),
    swaggerUi.setup(openapi_v2_doc),
  );

  // OpenAPI Validator
  app.use(
    OpenApiValidator.middleware({
      apiSpec: spec_path,
      validateRequests: true,
      validateResponses: true,
      validateSecurity: process.env["AUTHENTICATION_ENABLED"]!.toLowerCase() === "true",
    }),
  );

  app.use("/api/v2/", addDefaultProperties);
  app.use("/api/v2/ping", pingRouter);
  app.use("/api/v2/system", authenticateMiddleware, systemRouter);

  app.use(
    "/api/v2/authenticate",
    authenticationRouter(
      process.env["AUTHENTICATION_ENABLED"]!,
      parseInt(process.env["JWT_EXPIRATION"]!),
      process.env["JWT_SECRET"]!,
    ),
  );

  // The real data routes
  app.use("/api/v2/sensors", authenticateMiddleware, sensorsRouter);
  app.use("/api/v2/outputs", authenticateMiddleware, outputsRouter);
  app.use("/api/v2/automations", authenticateMiddleware, automationsRouter);
  app.use("/api/v2/output-actions", authenticateMiddleware, outputActionsRouter);
  app.use("/api/v2/notification-actions", authenticateMiddleware, notificationsRouter);
  app.use("/api/v2/camera", authenticateMiddleware, cameraRouter);
  app.use("/api/v2/subcontrollers", authenticateMiddleware, subcontrollersRouter);
  app.use("/api/v2/device-zones", authenticateMiddleware, deviceZoneRouter);
  app.use("/api/v2/journals", authenticateMiddleware, journalRouter);
  app.use("/api/v2/entries", authenticateMiddleware, entryRouter);
  app.use("/api/v2/tags", authenticateMiddleware, tagRouter);

  // Error handler - anything unexpected ends up here.
  app.use((err: any, req: Request, res: Response, _next: NextFunction) => {
    // format error
    const errorResponse = {
      statusCode: err.status ?? 500,
      error: {
        request: {
          method: req.method,
          url: req.originalUrl,
          body: req.body,
          query: req.query,
        },
        name: err.name ?? "Internal Server Error",
        url: req.originalUrl,
        details: err.errors ?? [],
      },
      ...(res.locals["defaultProperties"] ?? createDefaultProperties()),
    };

    if (err instanceof ContractValidationError) {
      const validationLogPayload = {
        operationId: err.operationId,
        phase: err.phase,
        method: req.method,
        url: req.originalUrl,
        details: err.errors,
      };

      if (
        process.env["NODE_ENV"] === "test" &&
        req.originalUrl.startsWith("/api/v2/outputs/chart-data")
      ) {
        console.error(
          "[chart-debug] contract validation error",
          JSON.stringify({
            operationId: err.operationId,
            phase: err.phase,
            method: req.method,
            url: req.originalUrl,
            details: err.errors,
            stack: err.stack,
          }),
        );
      }

      if (err.phase === "request") {
        logger.warn(`Contract validation failed: ${JSON.stringify(validationLogPayload)}`);
      } else {
        logger.error(`Contract validation failed: ${JSON.stringify(validationLogPayload)}`);
      }
    } else if (err.status === 500) {
      if (
        process.env["NODE_ENV"] === "test" &&
        req.originalUrl.startsWith("/api/v2/outputs/chart-data")
      ) {
        console.error(
          "[chart-debug] error middleware",
          JSON.stringify({
            method: req.method,
            url: req.originalUrl,
            name: err?.name,
            message: err?.message,
            status: err?.status ?? 500,
            stack: err?.stack,
            errors: err?.errors,
          }),
        );
      }
      logger.error(JSON.stringify(errorResponse));
    }

    res.status(err.status ?? 500).json(errorResponse);
  });
}

export default ApiRootV2;